requirements.txt
celery==5.3.0
kombu==5.3.0
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-instrumentation-celery==0.36b0
